+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix "quantity" showing up for variants
   - (label for "quantity to add to cart" which was hidden yesterday)
 - show for staff in product details the qty available for each variant
+- add `django.contrib.admin`
 
 
 ## 2018-10-01